+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/update/BotSubscriptionUpdatedUpdate.kt
@@ -1,0 +1,12 @@
+package dev.inmo.tgbotapi.types.update
+
+import dev.inmo.tgbotapi.types.UpdateId
+import dev.inmo.tgbotapi.types.payments.BotSubscriptionUpdated
+import dev.inmo.tgbotapi.types.update.abstracts.Update
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class BotSubscriptionUpdatedUpdate(
+    override val updateId: UpdateId,
+    override val data: BotSubscriptionUpdated
+) : Update