+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/update/RawUpdate.kt
@@ -15,6 +15,7 @@ import dev.inmo.tgbotapi.types.chat.member.ChatMemberUpdated
 import dev.inmo.tgbotapi.types.managed_bots.ManagedBotUpdated
 import dev.inmo.tgbotapi.types.message.abstracts.*
 import dev.inmo.tgbotapi.types.message.payments.PaidMediaPurchased
+import dev.inmo.tgbotapi.types.payments.BotSubscriptionUpdated
 import dev.inmo.tgbotapi.types.payments.PreCheckoutQuery
 import dev.inmo.tgbotapi.types.payments.ShippingQuery
 import dev.inmo.tgbotapi.types.polls.Poll
@@ -69,6 +70,7 @@ internal data class RawUpdate constructor(
     private val guest_message: RequestGuestContentMessage<*>? = null,
     private val purchased_paid_media: PaidMediaPurchased? = null,
     private val managed_bot: ManagedBotUpdated? = null,
+    private val subscription: BotSubscriptionUpdated? = null,
 ) {
     @Transient
     private var initedUpdate: Update? = null
@@ -107,6 +109,7 @@ internal data class RawUpdate constructor(
                 guest_message != null -> GuestMessageUpdate(updateId, guest_message)
                 purchased_paid_media != null -> PaidMediaPurchasedUpdate(updateId, purchased_paid_media)
                 managed_bot != null -> ManagedBotUpdate(updateId, managed_bot)
+                subscription != null -> BotSubscriptionUpdatedUpdate(updateId, subscription)
                 else -> UnknownUpdate(
                     updateId,
                     raw